fix: requestAlwaysAuthorization sometimes doesn't return
In some circumstances, requesting location permission has no effect. When this happens, the call to request permission never returns. This fix uses the UIApplication state as a proxy for knowing if an authorization dialog has been shown, and when it has been dismissed so we can return a value in the case that the dialog is never shown.

fixes #30